解决加密
解决加密

+++ KCKPLeader/Tool/NSString+NSStringMD5.m
@@ -15,7 +15,7 @@
 {
     const char *cStr = [self UTF8String];
     unsigned char digest[CC_MD5_DIGEST_LENGTH];
-    CC_MD5( cStr, strlen(cStr), digest );
+    CC_MD5( cStr,(unsigned int) strlen(cStr), digest );
     
     NSMutableString *output = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
